Move overlays into a separate folder
