+++ src/PixiEditor/ViewModels/SubViewModels/Tools/Tools/PenToolViewModel.cs
@@ -5,7 +5,7 @@ using PixiEditor.DrawingApi.Core.Numerics;
 using PixiEditor.Models.Commands.Attributes.Commands;
 using PixiEditor.ViewModels.SubViewModels.Tools.ToolSettings.Settings;
 using PixiEditor.ViewModels.SubViewModels.Tools.ToolSettings.Toolbars;
-using PixiEditor.Views.UserControls.BrushShapeOverlay;
+using PixiEditor.Views.UserControls.Overlays.BrushShapeOverlay;
 
 namespace PixiEditor.ViewModels.SubViewModels.Tools.Tools
 {
