--- src/PixiEditor/Views/UserControls/BrushShapeOverlay/BrushShape.cs
@@ -1,8 +1,0 @@
-﻿namespace PixiEditor.Views.UserControls.BrushShapeOverlay;
-internal enum BrushShape
-{
-    Hidden,
-    Pixel,
-    Square,
-    Circle
-}
