+++ src/PixiEditor/Views/UserControls/Overlays/BrushShapeOverlay/BrushShape.cs
@@ -1,0 +1,13 @@
+﻿using PixiEditor;
+using PixiEditor.Views;
+using PixiEditor.Views.UserControls;
+using PixiEditor.Views.UserControls.Overlays.BrushShapeOverlay;
+
+namespace PixiEditor.Views.UserControls.Overlays.BrushShapeOverlay;
+internal enum BrushShape
+{
+    Hidden,
+    Pixel,
+    Square,
+    Circle
+}
