+++ src/PixiEditor/Views/UserControls/Overlays/BrushShapeOverlay/BrushShapeOverlay.cs
@@ -4,9 +4,13 @@ using System.Windows.Input;
 using System.Windows.Media;
 using ChunkyImageLib.DataHolders;
 using ChunkyImageLib.Operations;
+using PixiEditor;
 using PixiEditor.DrawingApi.Core.Numerics;
+using PixiEditor.Views;
+using PixiEditor.Views.UserControls;
+using PixiEditor.Views.UserControls.Overlays.BrushShapeOverlay;
 
-namespace PixiEditor.Views.UserControls.BrushShapeOverlay;
+namespace PixiEditor.Views.UserControls.Overlays.BrushShapeOverlay;
 #nullable enable
 internal class BrushShapeOverlay : Control
 {
